pyproject.toml
[project]
name = "wiki-pt-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chonkie[semantic]>=0.4.1",
    "datasets>=3.2.0",
    "duckdb>=1.2.0",
    "polars>=1.21.0",
    "python-dotenv>=1.0.1",
    "sentence-transformers>=3.4.0",
]
